refactor get current user, integration

+++ src/main/java/com/mgmtp/radio/controller/v1/FavoriteSongController.java
@@ -4,6 +4,8 @@ import com.mgmtp.radio.controller.BaseRadioController;
 import com.mgmtp.radio.controller.response.RadioSuccessResponse;
 import com.mgmtp.radio.dto.user.FavoriteSongDTO;
 import com.mgmtp.radio.exception.RadioBadRequestException;
+import com.mgmtp.radio.exception.RadioException;
+import com.mgmtp.radio.exception.RadioNotFoundException;
 import com.mgmtp.radio.service.user.FavoriteSongService;
 import com.mgmtp.radio.support.validator.user.CreateFavoriteSongValidator;
 import lombok.extern.log4j.Log4j2;
@@ -37,26 +39,43 @@ public class FavoriteSongController extends BaseRadioController {
 
 	@GetMapping
 	@ResponseStatus(HttpStatus.OK)
-	public Flux<RadioSuccessResponse<FavoriteSongDTO>> getAll() {
-		String userId = getCurrentUser().getId();
-		return favoriteSongService.findByUserId(userId).map(RadioSuccessResponse::new);
+	public Flux<RadioSuccessResponse<FavoriteSongDTO>> getAll() throws RadioException {
+
+		if(getCurrentUser().isPresent()) {
+		    String userId = getCurrentUser().get().getId();
+			return favoriteSongService.findByUserId(userId).map(RadioSuccessResponse::new);
+		} else {
+			throw new RadioNotFoundException("unauthorized");
+		}
 	}
 
 
 	@PostMapping
 	@ResponseStatus(HttpStatus.OK)
-	public Mono<RadioSuccessResponse<FavoriteSongDTO>> store(@Validated @RequestBody FavoriteSongDTO favoriteSongDTO, BindingResult bindingResult) {
+	public Mono<RadioSuccessResponse<FavoriteSongDTO>> store(@Validated @RequestBody FavoriteSongDTO favoriteSongDTO,
+                                                             BindingResult bindingResult)
+            throws RadioException {
 		if (bindingResult.hasErrors()) {
 			return Mono.error(new RadioBadRequestException(bindingResult.getAllErrors().get(0).getDefaultMessage()));
 		}
-		String userId = getCurrentUser().getId();
-		return favoriteSongService.create(userId, favoriteSongDTO).map(RadioSuccessResponse::new);
+
+        if(getCurrentUser().isPresent()) {
+            String userId = getCurrentUser().get().getId();
+            return favoriteSongService.create(userId, favoriteSongDTO).map(RadioSuccessResponse::new);
+        } else {
+            throw new RadioNotFoundException("unauthorized");
+        }
 	}
 
 	@DeleteMapping("/{id}")
 	@ResponseStatus(HttpStatus.OK)
-	public Mono<RadioSuccessResponse<Void>> delete(@PathVariable(value = "id") String id) {
-		String userId = getCurrentUser().getId();
-		return favoriteSongService.delete(id, userId).map(song -> new RadioSuccessResponse<>());
+	public Mono<RadioSuccessResponse<Void>> delete(@PathVariable(value = "id") String id) throws RadioException {
+        if(getCurrentUser().isPresent()) {
+            String userId = getCurrentUser().get().getId();
+            return favoriteSongService.delete(id, userId).map(song -> new RadioSuccessResponse<>());
+        } else {
+            throw new RadioNotFoundException("unauthorized");
+        }
+
 	}
 }
